Suppor for single char bool support
The PyYAML implementation doesn't correctly decode single char bool
variants 'y/Y' and 'n/N'. This breaks compatibility with parsers that
correctly implement the YAML spec.

HOWEVER, this change dramatically impacts users, who now need to quote
keys 'y' and 'n'. E.g.

``` yaml
op:
  flags:
    x: 123
    'y': 456
```

If `y` in this case isn't quoted, it's decoded as Python `True`, which
is not the user intent.

So we cannot accept this change and will immediately revert it. But
the change and revert ought to be a part of the Git history.