Include `config` sources in exempted source code files
A `config` file is just another project local file. It's different
from `file` in that Guild modifies its contents to reflect the run
flag values. If this file is copied as source code, the default
resolution process will not overwrite it (correctly). We need to
ignore this file in source code copy to reflect the intent to use the
file as a dependency.