Bring scikit-optimize into code base as `guild._skopt`
[scikit-optimize](https://github.com/scikit-optimize/scikit-optimize)
is lightly maintained and the current version as of this commit (0.9)
is incompatible with the latest versions of both `scikit-learn` and
`numpy`. While we would prefer to patch libraries like this to ensure
both forward and backward compatibility with other libraries, the
patch requirements in this case would be brittle, requiring wholesale
duplication of large blocks of code.

This commit is an exact copy of the tag `v0.9.0` and is not intended
to work as a replacement. Subsequent commits will apply the required
changes to Guild to incorporate the now internal library.