Cleanup uat
- Fix tests in general
- Consolidate multiple remote tests

uat suffers from test interdepepdencies (sequential) which makes
fixing them challenigng as broken tests, which need to be fixed, may
rely on changes to previous tests that have already passed and need to
be re-run.

Future uat tests should be written to be independent of one
another. Further consolidation of uat docs may be done as needed.