Set core count to 4 for timing critical tests
This is a bit of a hail mary to introduce some consistency in test
results across platforms for timing critical tests. What this test
flag effectively does is skip the timing critial tests in the CI env,
which provides palty compute environments (2 and 3 core virtualized
systems).

These tests will run on developer workstations and any decent hardware
we end up getting access to in CI envs.