Set target-path on config flags interface
Guild was writing resolved config to the run root, even when the
source file is located in a project subdir. While the user can always
configure a dep source to explicitly set the target, Guild was making
the wrong assumption in this case -- the user is far more likley to
intend the resolved file to be saved to the same relative path than to
the run root.