Bump timeouts/waits for make check target
`make check` is using a concurrency of 8, which can cause false test
failures on some systems due to load. Bumping timeout/wait values here
is an attempt to address that on developer systems.

In cases where higher timtouts are needed (in CI env, a different set
of env vars may be needed. `Makefile` should not be changed in these
cases.