Fix #518
Allows an operation to explicitly disable plugins by setting `plugins`
to an empty list or to false.

Note that the docs are still incorrect when they say that plugins are
not enabled by default. Plugins most certainly are enabled by deault -
that's central to Guild's design. However, operations must be able to
fully specify the plugins that are enabled when it's run, which this
commit addresses.

See `guild/tests/plugins.md` for details.